refactor(plugin): 收束公共协议与清理依赖方向
