+++ packages/model/lib/model/index.ts
@@ -4,5 +4,6 @@ export * from './download'
 export * from './ep'
 export * from './image'
 export * from './item'
+export * from './plugin'
 export * from './resource'
 export * from './user'
