+++ packages/plugin/lib/api/config.ts
@@ -1,0 +1,21 @@
+import type { FormSingleConfigure } from '@delta-comic/model'
+
+export type ConfigDescription = Record<
+  string,
+  Required<Pick<FormSingleConfigure, 'defaultValue'>> & FormSingleConfigure
+>
+
+export type UnwrapConfigPointer<T extends ConfigPointer> = T['_type']
+
+export class ConfigPointer<T extends ConfigDescription = ConfigDescription> {
+  public readonly key: symbol
+  public readonly _type = {} as T
+
+  public constructor(
+    public readonly pluginName: string,
+    public readonly config: T,
+    public readonly configName: string,
+  ) {
+    this.key = Symbol.for(`config:${pluginName}`)
+  }
+}
