+++ packages/plugin/lib/api/env.ts
@@ -1,0 +1,6 @@
+export type Platform = 'tauri' | 'web'
+
+export interface ConfigEnv {
+  safe: boolean
+  platform: Platform
+}
