+++ packages/plugin/lib/api/hook/content.ts
