+++ packages/plugin/lib/api/hook/lifecycle.ts
