+++ packages/plugin/lib/api/i18n.ts
@@ -1,0 +1,5 @@
+export interface PluginLocaleMessage {
+  [key: string]: PluginLocaleMessage | string
+}
+
+export type PluginLocaleMessages = Record<string, PluginLocaleMessage>
