+++ packages/plugin/lib/core/index.ts
@@ -1,5 +1,4 @@
-import { defineDeltaComicPlugin } from '../export'
-import { useConfig } from '../module/config/store'
+import { defineDeltaComicPlugin } from '../api'
 
 import { cfg } from './config'
 export { cfg } from './config'
@@ -8,13 +7,7 @@ import { pluginName } from './env'
 import { tokenInit, nativeInit, tokenShare } from './share'
 
 export default defineDeltaComicPlugin(() => ({
+  config: cfg,
   name: pluginName,
-  i18nName: pluginName,
-  hooks: {
-    onPreboot() {
-      const cs = useConfig()
-      cs.$register(cfg)
-    },
-  },
   model: { social: { share: { initiative: [tokenInit, nativeInit], tokenListen: [tokenShare] } } },
 }))
