+++ packages/plugin/lib/index.ts
@@ -1,4 +1,5 @@
-export * from './export'
+export * from './api'
+export * from './kernel'
 export * from './module'
 export * from './global'
